Revert moving tets to Old directory
