+++ .gitignore
@@ -5,4 +5,4 @@ composer.phar
 composer.lock
 runtests.sh
 /ssh.*
-/tests/Old/data/out/tables/
+/tests/data/out/tables/
