+++ docker-compose.yml
@@ -25,7 +25,7 @@ services:
       MYSQL_ROOT_PASSWORD: somePassword
 
   sshproxy:
-    build: ./tests/Old/env/sshproxy
+    build: ./tests/env/sshproxy
     command: sh -c 'echo $SSH_KEY_PUBLIC >> ~/.ssh/authorized_keys && /usr/sbin/sshd -D'
     environment:
       - SSH_KEY_PUBLIC
